third_party/WebKit/LayoutTests/external/wpt/IndexedDB/idbobjectstore_createIndex15-autoincrement-expected.txt

This is a testharness.js-based test.
PASS Explicit Primary Key
FAIL Auto-Increment Primary Key assert_equals: Expected 100. expected (number) 100 but got (object) null
Harness: the test ran to completion.
